リファクタリングした
