+++ reactjs/src/connects/ProfileConnect.tsx
@@ -1,9 +1,6 @@
 import { connect } from 'react-redux';
 import { bindActionCreators, Dispatch } from 'redux';
-import shortId from 'shortid';
 import { fetchUserWithShortId } from '~/actions/userAction';
-
-import firebase from '~/lib/firebase';
 
 const mapDispatchToProps = (dispatch: Dispatch) =>
   bindActionCreators(
@@ -21,42 +18,6 @@ const mapStateToProps = (state: any, props: any) => {
   const isLogin = !profile.isEmpty && profile.isLoaded;
   const isMyProfile = profile.shortId === userShortId;
 
-  const myResourceId = user ? `/users/${user.id}` : null;
-
-  const handleSensitiveListUpdate = (type: 'mute' | 'block') => () => {
-    const newData = {
-      id: shortId.generate(),
-      createdAt: new Date(),
-      userId: user.id,
-      userShortId: user.shortId,
-      userDisplayName: user.displayName,
-      userPhotoURL: user.photoURL
-    };
-
-    const updateData =
-      type === 'mute'
-        ? Object.assign(
-            {
-              muteList: firebase.firestore.FieldValue.arrayUnion(newData)
-            },
-            { createdAt: new Date() }
-          )
-        : Object.assign(
-            {
-              blockList: firebase.firestore.FieldValue.arrayUnion(newData)
-            },
-            { createdAt: new Date() }
-          );
-
-    return (
-      myResourceId &&
-      firebase
-        .firestore()
-        .doc(myResourceId)
-        .update(updateData)
-    );
-  };
-
   return {
     user,
     loading: state.user.loading,
@@ -64,7 +25,6 @@ const mapStateToProps = (state: any, props: any) => {
     userShortId,
     isLogin,
     isMyProfile,
-    handleSensitiveListUpdate,
     ...props
   };
 };
