+++ reactjs/src/web/containers/BlockButtonContainer.tsx
@@ -1,0 +1,4 @@
+import connect from '~/connects/BlockButtonConnect';
+import BlockButton from '~/web/components/atoms/BlockButton';
+
+export default connect(BlockButton);
