+++ reactjs/src/web/containers/MuteButtonContainer.tsx
@@ -1,0 +1,4 @@
+import connect from '~/connects/MuteButtonConnect';
+import MuteButton from '~/web/components/atoms/MuteButton';
+
+export default connect(MuteButton);
